backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
pydantic==2.9.2
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
bcrypt==5.0.0
python-multipart==0.0.12
alembic==1.14.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
email-validator==2.3.0
ollama==0.3.1

